+++ docs/changelog.rst
@@ -1,6 +1,20 @@
 Changelog 📝
 ============
 
+`v2021.5.0 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v2021.5.0>`_ (2021-05-05)
+-------------------------------------------------------------------------------------
+**New Features:**
+
+- `#418 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418>`_: ``compute vm set retire mk``: manage retirement requests for vms.
+- `#418 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418>`_: ``compute vm set retire confirm``: manage retirement requests for vms.
+- `#418 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418>`_: ``compute vm set retire cancel``: manage retirement requests for vms.
+- `#418 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/418>`_: ``compute vm set retire send``: manage retirement requests for vms.
+- `#419 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/419>`_: ``compute vm get retire``: get retirement requests for vm.
+
+**Improvements:**
+
+- `#417 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/issues/417>`_: ``core``: move from ``semver`` to ``calver``.
+
 `v0.12.1 <https://gitlab-ee.eis.utoronto.ca/vss/vss-cli/tags/v0.12.1>`_ (2021-04-15)
 -------------------------------------------------------------------------------------